BMI Calculator
This repository just contains the node.js and express.js code required to design the calculator. 
Besides, we need to have installed node modules to run this project on our local server.
The file named "Use of Command Line" contains all the commands that were used to install different npm packages and run our local server on the port - 3090.